+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 - Add `MntFlags` and `unmount` on all of the BSDs.
   ([#1849](https://github.com/nix-rust/nix/pull/1849))
+- Added a 'Statfs::flags' method.
+  ([#1849](https://github.com/nix-rust/nix/pull/1849))
 - Added `NSFS_MAGIC` FsType on Linux and Android.
   ([#1829](https://github.com/nix-rust/nix/pull/1829))
 - Added `sched_getcpu` on platforms that support it.